ASPACE-300 refactor BatchProcessor - create JobExecutionManager and JobResultProcessor classes to handle running FOLIO jobs and updating db based on the job result
